requirements.txt
# ==========================================
# CORE
# ==========================================
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.0

# ==========================================
# DATABASE
# ==========================================
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
greenlet==3.1.0

# ==========================================
# REDIS
# ==========================================
redis[hiredis]==5.1.0

# ==========================================
# HTTP CLIENT (para llamadas a APIs)
# ==========================================
httpx==0.27.2
aiohttp==3.10.5

# ==========================================
# GOOGLE (Gemini + Calendar)
# ==========================================
google-generativeai==0.8.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

# ==========================================
# UTILIDADES
# ==========================================
python-multipart==0.0.9
python-dotenv==1.0.1
apscheduler==3.10.4

# ==========================================
# DESARROLLO
# ==========================================
pytest==8.3.3
pytest-asyncio==0.24.0
